fix install issue and add github pubkey steps

+++ install.sh
@@ -4,7 +4,7 @@ echo "Setting up your Mac..."
 
 # Install homebrew
 echo "Installing homebrew..."
-if test ! $(which brew); then
+if [ ! -f $(which brew) ]; then
   /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
 fi
 
@@ -17,10 +17,10 @@ echo "Installing bundles..."
 brew tap homebrew/bundle
 brew bundle
 
+# Create links
 echo "Creating essential file links..."
 # Create symlink and restart shell
-if [ ! -f ~/.hammerspoon ]
-then
+if [ -d ~/.hammerspoon ]; then
 	rm -rf ~/.hammerspoon
 fi
 ln -s ~/.dotfiles/hammerspoon ~/.hammerspoon
@@ -28,22 +28,21 @@ ln -sf ~/.dotfiles/.gitignore ~/.gitignore
 ln -sf ~/.dotfiles/.gitconfig ~/.gitconfig
 ln -sf ~/.dotfiles/.editorconfig ~/.editorconfig
 ln -sf ~/.dotfiles/.hyper.js ~/.hyper.js
-#ln -sf ~/.dotfiles/.mackup.cfg ~/.mackup.cfg
 ln -sf ~/.dotfiles/.zshrc ~/.zshrc
 
-# Set macOS preferences
-# We will run this last because this will reload the shell
-# source .macos
+# Get pubkey from github
+echo "Getting pubkey from github..."
+mkdir -p ~/.ssh
+# if id_rsa.pub doesn't exist
+if [ ! -f ~/.ssh/id_rsa.pub ]; then
+    curl https://github.com/gddabe.keys | tee -a ~/.ssh/id_rsa.pub
+fi
 
 # Dock cleanup
 echo "Reseting the dock..."
 dockutil --no-restart --remove all
 dockutil --no-restart --add "/Applications/Google Chrome.app"
 dockutil --no-restart --add "/Applications/Hyper.app"
-
-echo "Getting pubkey from github..."
-mkdir -p ~/.ssh
-curl https://github.com/gddabe.keys | tee -a ~/.ssh/id_rsa.pub
 
 killall Dock
